Massive update
Tons of changes. I'm moving back (again) to a single repo for both bash and zsh. The thinking being that since both shells are compatible (for the most part) with one another I might as
well re-use as much code and logic as I can with my aliases and functions. Also, I'm moving away from the "oh-my-zsh" or "antigen" approach to zsh. (It's just way to bulky)

Still very much work in progress. Stay tuned...